Refactored array job submission, added support for flags.
Array jobs are now specified when calling submit() via an optional
argument instead of the nodes property.

Added support for flags with no associated value (e.g. --test
and --spread-job).  These values are specified with an empty string
as their value.  This likely needs a refactor, but it's a viable
workaround for the moment.